+++ src/App.js
@@ -1,5 +1,5 @@
-import logo from './logo.svg';
-import './App.css';
+import logo from './resources/images/logo.svg';
+import './resources/css/App.css';
 
 function App() {
   return (
